pulling out shared methods for extension registry test cases into a mixin class
